refactor: [SCRUM-159] mybatis to JPA repo

+++ src/main/java/com/dottree/nonogrammers/repository/MyPageRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MyPageRepository extends JpaRepository<User, Long> {
+public interface MyPageRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findByNickName(String nickName);
 }
